Move LoginScreen.kt to new auth/ package
